Separating out admin.py, adding challenge types
