+++ CTFd/__init__.py
@@ -68,7 +68,7 @@ def create_app(config='CTFd.config.Config'):
         from CTFd.challenges import challenges
         from CTFd.scoreboard import scoreboard
         from CTFd.auth import auth
-        from CTFd.admin import admin
+        from CTFd.admin import admin, admin_statistics, admin_challenges, admin_pages, admin_scoreboard, admin_containers, admin_keys, admin_teams
         from CTFd.utils import init_utils, init_errors, init_logs
 
         init_utils(app)
@@ -79,7 +79,15 @@ def create_app(config='CTFd.config.Config'):
         app.register_blueprint(challenges)
         app.register_blueprint(scoreboard)
         app.register_blueprint(auth)
+
         app.register_blueprint(admin)
+        app.register_blueprint(admin_statistics)
+        app.register_blueprint(admin_challenges)
+        app.register_blueprint(admin_teams)
+        app.register_blueprint(admin_scoreboard)
+        app.register_blueprint(admin_keys)
+        app.register_blueprint(admin_containers)
+        app.register_blueprint(admin_pages)
 
         from CTFd.plugins import init_plugins
