+++ CTFd/static/admin/js/templates/challenges/standard/standard-challenge-create.hbs
@@ -1,0 +1,79 @@
+<div class="col-md-6 col-md-offset-3">
+    <form method="POST" action="{{ request.script_root }}/admin/chal/new" enctype="multipart/form-data">
+        <div class="form-group">
+            <label for="name">Name</label>
+            <input type="text" class="form-control" name="name" placeholder="Enter challenge name">
+        </div>
+        <div class="form-group">
+            <label for="category">Category</label>
+            <input type="text" class="form-control" name="category" placeholder="Enter challenge category">
+        </div>
+
+        <ul class="nav nav-tabs" role="tablist" id="new-desc-edit">
+            <li role="presentation" class="active"><a href="#new-desc-write" aria-controls="home" role="tab" data-toggle="tab">Write</a></li>
+            <li role="presentation"><a href="#new-desc-preview" aria-controls="home" role="tab" data-toggle="tab">Preview</a></li>
+        </ul>
+        <div class="tab-content">
+            <div role="tabpanel" class="tab-pane active" id="new-desc-write">
+                <div class="form-group">
+                    <label for="message-text" class="control-label">Message:</label>
+                    <textarea id="new-desc-editor" class="form-control" name="desc" rows="10"></textarea>
+                </div>
+            </div>
+            <div role="tabpanel" class="tab-pane content" id="new-desc-preview">
+            </div>
+        </div>
+
+        <div class="form-group">
+            <label for="value">Value</label>
+            <input type="number" class="form-control" name="value" placeholder="Enter value" required>
+        </div>
+        <div class="row">
+            <div class="form-group">
+                <div class="col-md-8">
+                    <label for="key">Key</label>
+                    <input type="text" class="form-control" name="key" placeholder="Enter key">
+                </div>
+                <div class="col-md-2">
+                    <label class="radio-inline">
+                        <input type="radio" name="key_type[0]" value="0" checked>
+                        Static
+                    </label>
+                </div>
+                <div class="col-md-2">
+                    <label class="radio-inline">
+                        <input type="radio" name="key_type[0]" value="1"> Regex
+                    </label>
+                </div>
+            </div>
+        </div>
+
+        <div class="row">
+            <div class="form-group">
+                <div class="col-md-9">
+                    <div class="checkbox">
+                        <label>
+                            <input name="hidden" type="checkbox">
+                            Hide challenge on creation
+                        </label>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <div class="row">
+            <div class="form-group">
+                <div class="col-md-9">
+                    <label>Upload challenge files</label>
+                    <sub class="help-block">Attach multiple files using Control+Click or Cmd+Click.</sub>
+                    <input type="file" name="files[]" multiple="multiple">
+                </div>
+            </div>
+        </div>
+        <input type="hidden" value="{{ nonce }}" name="nonce" id="nonce">
+        <input type="hidden" value="0" name="chaltype" id="chaltype">
+        <div style="text-align:center">
+            <button class="btn btn-theme btn-outlined create-challenge-submit" type="submit">Create</button>
+        </div>
+    </form>
+</div>
