+++ CTFd/static/admin/js/templates/keys/regex/edit-regex-modal.hbs
