+++ CTFd/static/admin/js/templates/keys/regex/regex.hbs
