+++ CTFd/static/admin/js/templates/keys/static/edit-static-modal.hbs
