+++ CTFd/static/admin/js/templates/keys/static/static.hbs
