+++ CTFd/static/original/js/templates/challenges/standard/standard-challenge-script.js
@@ -1,0 +1,3 @@
+$('#submit-key').click(function (e) {
+    submitkey($('#chal-id').val(), $('#answer-input').val(), $('#nonce').val())
+});
