+++ CTFd/templates/admin/chals.html
@@ -319,7 +319,7 @@
     <div style="text-align:center">
         <br>
         <h1 class="text-center">Challenges</h1>
-        <button class="btn btn-theme btn-outlined create-challenge">New Challenge</button>
+        <a class="btn btn-theme btn-outlined create-challenge" href="{{ request.script_root }}/admin/chal/new" role="button">New Challenge</a>
         <div>
             <table id='challenges'>
             </table>
