+++ CTFd/templates/original/base.html
@@ -14,6 +14,7 @@
     <link rel="stylesheet" type="text/css" href="{{ request.script_root }}/static/user.css">
     {% block stylesheets %}{% endblock %}
     <script src="{{ request.script_root }}/static/{{ ctf_theme() }}/js/vendor/moment.min.js"></script>
+    <script src="{{ request.script_root }}/static/original/js/vendor/handlebars.min.js"></script>
     <script type="text/javascript">
         var script_root = "{{ request.script_root }}";
     </script>
